+++ tasks/slacktest.ts
@@ -1,8 +1,67 @@
-import { danger } from "danger";
+import { danger, peril } from "danger";
+import { IncomingWebhook, IncomingWebhookSendArguments } from "@slack/client";
 
 const org = "gatsbyjs";
 const label = "stale?";
 
+/**
+ * A task that accepts Slack incoming webhook data
+ * and sends a message into the Artsy Dev chat room.
+ *
+ * The full API docs for the syntax of the expected data
+ * can be found: https://slackapi.github.io/node-slack-sdk/reference/IncomingWebhook
+ *
+ * Usage in a Dangerfile:
+ *
+    const message = {
+      unfurl_links: false,
+      attachments: [
+        {
+          pretext: "We can throw words around like two hundred million galaxies",
+          color: "good",
+          title: issue.title,
+          title_link: issue.html_url,
+          author_name: issue.user.login,
+          author_icon: issue.user.avatar_url,
+        },
+      ],
+    }
+    peril.runTask("slack-dev-channel", "in 5 minutes", message)
+ */
+
+/**
+ * The default, send a slack message with some data that's come in
+ * this is also usable as a task
+ */
+
+const slackData = async (data: IncomingWebhookSendArguments) => {
+  if (!data) {
+    console.log(
+      "No data was passed to slack-dev-channel, so a message will not be sent."
+    );
+  } else {
+    const url = peril.env.SLACK_WEBHOOK_URL || "";
+    const webhook = new IncomingWebhook(url);
+    await webhook.send(data);
+  }
+};
+
+/**
+ * Send a slack message to the dev channel in Artsy
+ * @param message the message to send to #dev
+ */
+const slackMessage = async (message: string) => {
+  const data = {
+    unfurl_links: false,
+    attachments: [
+      {
+        color: "good",
+        title: message
+      }
+    ]
+  };
+  await slackData(data);
+};
 export interface Result {
   url: string;
   repository_url: string;
@@ -31,8 +90,6 @@ export interface Result {
 // https://developer.github.com/v3/search/#search-issues
 
 export default async () => {
-  const { slackMessage, slackData } = await import("./slackDevChannel");
-
   const api = danger.github.api;
   const staleQuery = `org:${org} label:${label} state:open`;
   const searchResponse = await api.search.issues({ q: staleQuery });
